@@ -72,91 +75,88 @@
     "react-motion": "^0.5.2",
     "react-redux": "^4.x.x",
     "redux": "^3.x.x",
-    "redux-immutable": "3.0.8",
-    "remarkable": "^1.7.1",
+    "redux-immutable": "3.1.0",
+    "remarkable": "^1.7.4",
     "reselect": "^2.5.4",
     "serialize-error": "^2.1.0",
-    "swagger-client": "^3.8.22",
-    "url-parse": "^1.1.8",
+    "sha.js": "^2.4.11",
+    "swagger-client": "^3.10.0",
+    "url-parse": "^1.4.7",
     "xml-but-prettier": "^1.0.1",
     "zenscroll": "^4.0.2"
   },
   "devDependencies": {
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.5.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.4.4",
+    "@release-it/conventional-changelog": "^1.1.0",
     "autoprefixer": "^8.4.1",
-    "babel-core": "^6.23.1",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^7.1.0",
-    "babel-plugin-module-alias": "^1.6.0",
-    "babel-plugin-transform-es2015-constants": "^6.1.4",
+    "babel-eslint": "^9.0.0",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-module-resolver": "^3.2.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.13",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-es2015-ie": "^6.6.2",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-0": "^6.22.0",
-    "babel-runtime": "^6.23.0",
-    "bundlesize": "^0.17.0",
+    "body-parser": "^1.19.0",
+    "bundlesize": "^0.18.0",
     "chromedriver": "^2.38.3",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.11",
-    "cypress": "^3.1.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "cors": "^2.8.5",
+    "css-loader": "^3.1.0",
+    "cypress": "^3.4.1",
     "dedent": "^0.7.0",
     "deepmerge": "^2.1.0",
     "enzyme": "^2.7.1",
     "eslint": "^4.1.1",
-    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-mocha": "^4.11.0",
-    "eslint-plugin-react": "^7.10.0",
+    "eslint-plugin-react": "^7.14.3",
     "expect": "^1.20.2",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.11",
-    "git-describe": "^4.0.1",
+    "express": "^4.17.1",
+    "file-loader": "^4.1.0",
+    "git-describe": "^4.0.4",
     "http-server": "^0.11.1",
+    "ignore-assets-webpack-plugin": "^2.0.1",
     "imports-loader": "^0.8.0",
     "jsdom": "^11.10.0",
     "json-loader": "^0.5.7",
-    "json-server": "^0.12.2",
-    "less": "^3.0.2",
+    "json-merger": "^1.1.1",
+    "json-server": "^0.15.0",
+    "less": "^3.9.0",
     "license-checker": "^19.0.0",
+    "mini-css-extract-plugin": "^0.8.0",
     "mocha": "^5.1.1",
-    "nightwatch": "^0.9.16",
-    "node-sass": "^4.5.0",
-    "npm-run-all": "^4.1.2",
-    "null-loader": "0.1.1",
-    "nyc": "^11.3.0",
-    "open": "0.0.5",
-    "postcss-loader": "^2.1.5",
-    "raw-loader": "0.5.1",
+    "nightwatch": "^1.1.13",
+    "node-sass": "^4.13.1",
+    "npm-audit-ci-wrapper": "^2.3.0",
+    "npm-run-all": "^4.1.5",
+    "oauth2-server": "^2.4.1",
+    "open": "^6.4.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "postcss-loader": "^3.0.0",
+    "prettier": "^1.18.2",
+    "raw-loader": "^3.1.0",
     "react-test-renderer": "^15.5.4",
-    "release-it": "^7.6.1",
-    "rimraf": "^2.6.0",
-    "sass-loader": "^7.0.1",
-    "selenium-server-standalone-jar": "3.12.0",
+    "release-it": "^12.3.4",
+    "rimraf": "^2.6.3",
+    "sass-loader": "^7.1.0",
+    "selenium-server-standalone-jar": "^3.141.5",
+    "source-map-support": "^0.5.12",
     "standard": "^11.0.1",
-    "standard-loader": "^6.0.1",
-    "style-loader": "^0.21.0",
-    "tachyons-sass": "^4.9.2",
-    "uglifyjs-webpack-plugin": "^1.2.5",
-    "url-loader": "^1.0.1",
-    "webpack": "^3.1.0",
+    "tachyons-sass": "^4.9.5",
+    "terser-webpack-plugin": "^1.4.1",
+    "url-loader": "^2.1.0",
+    "webpack": "^4.39.0",
     "webpack-bundle-size-analyzer": "^2.5.0",
-    "webpack-cli": "^2.0.4",
-    "webpack-dev-server": "^2.11.1",
-    "worker-loader": "^1.1.1"
+    "webpack-cli": "^3.3.6",
+    "webpack-dev-server": "^3.7.2"
   },
   "config": {
     "deps_check_dir": ".deps_check"
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "IE 11"
-  ],
-  "nyc": {
-    "all": true,
-    "include": [
-      "**/src/core/plugins/**.js"
-    ]
   },
   "bundlesize": [
     {
