@@ -5,16 +5,16 @@ import Im from "immutable"
 const propStyle = { color: "#999", fontStyle: "italic" }
 
 export default class Headers extends React.Component {
-
   static propTypes = {
     headers: PropTypes.object.isRequired,
     getComponent: PropTypes.func.isRequired
   };
 
   render() {
-
     let { headers, getComponent } = this.props
+
     const Property = getComponent("Property")
+    const Markdown = getComponent("Markdown")
 
     if ( !headers || !headers.size )
       return null
@@ -36,12 +36,16 @@ export default class Headers extends React.Component {
               if(!Im.Map.isMap(header)) {
                 return null
               }
+              
+              const description = header.get("description")
               const type = header.getIn(["schema"]) ? header.getIn(["schema", "type"]) : header.getIn(["type"])
               const schemaExample = header.getIn(["schema", "example"])
 
               return (<tr key={ key }>
                 <td className="header-col">{ key }</td>
-                <td className="header-col">{ header.get( "description" ) }</td>
+                <td className="header-col">{
+                  !description ? null : <Markdown source={ description } />
+                }</td>
                 <td className="header-col">{ type } { schemaExample ? <Property propKey={ "Example" } propVal={ schemaExample } propStyle={ propStyle } /> : null }</td>
               </tr>)
             }).toArray()
