@@ -13,7 +13,7 @@ export default class BaseLayout extends React.Component {
   }
 
   render() {
-    let {specSelectors, getComponent} = this.props
+    let {errSelectors, specSelectors, getComponent} = this.props
 
     let SvgAssets = getComponent("SvgAssets")
     let InfoContainer = getComponent("InfoContainer", true)
@@ -22,9 +22,9 @@ export default class BaseLayout extends React.Component {
     let Models = getComponent("Models", true)
     let Row = getComponent("Row")
     let Col = getComponent("Col")
-    let ServersContainer = getComponent("ServersContainer", true)
     let Errors = getComponent("errors", true)
 
+    const ServersContainer = getComponent("ServersContainer", true)
     const SchemesContainer = getComponent("SchemesContainer", true)
     const AuthorizeBtnContainer = getComponent("AuthorizeBtnContainer", true)
     const FilterContainer = getComponent("FilterContainer", true)
@@ -33,21 +33,56 @@ export default class BaseLayout extends React.Component {
 
     const isSpecEmpty = !specSelectors.specStr()
 
-    if(isSpecEmpty) {
-      let loadingMessage
-      let isLoading = specSelectors.loadingStatus() === "loading"
-      if(isLoading) {
-        loadingMessage = <div className="loading"></div>
-      } else {
-        loadingMessage = <h4>No API definition provided.</h4>
-      }
+    const loadingStatus = specSelectors.loadingStatus()
 
+    let loadingMessage = null
+  
+    if(loadingStatus === "loading") {
+      loadingMessage = <div className="info">
+        <div className="loading-container">
+          <div className="loading"></div>
+        </div>
+      </div>
+    }
+
+    if(loadingStatus === "failed") {
+      loadingMessage = <div className="info">
+        <div className="loading-container">
+          <h4 className="title">Failed to load API definition.</h4>
+          <Errors />
+        </div>
+      </div>
+    }
+
+    if (loadingStatus === "failedConfig") {
+      const lastErr = errSelectors.lastError()
+      const lastErrMsg = lastErr ? lastErr.get("message") : ""
+      loadingMessage = <div className="info" style={{ maxWidth: "880px", marginLeft: "auto", marginRight: "auto", textAlign: "center" }}>
+        <div className="loading-container">
+          <h4 className="title">Failed to load remote configuration.</h4>
+          <p>{lastErrMsg}</p>
+        </div>
+      </div>
+    }
+
+    if(!loadingMessage && isSpecEmpty) {
+      loadingMessage = <h4>No API definition provided.</h4>
+    }
+
+    if(loadingMessage) {
       return <div className="swagger-ui">
         <div className="loading-container">
           {loadingMessage}
         </div>
       </div>
     }
+
+    const servers = specSelectors.servers()
+    const schemes = specSelectors.schemes()
+
+    const hasServers = servers && servers.size
+    const hasSchemes = schemes && schemes.size
+    const hasSecurityDefinitions = !!specSelectors.securityDefinitions()
 
     return (
 
@@ -61,15 +96,15 @@ export default class BaseLayout extends React.Component {
               </Col>
             </Row>
 
-            <div>
+            {hasServers || hasSchemes || hasSecurityDefinitions ? (
               <div className="scheme-container">
                 <Col className="schemes wrapper" mobile={12}>
-                  <ServersContainer />
-                  <SchemesContainer />
-                  <AuthorizeBtnContainer />
+                  {hasServers ? (<ServersContainer />) : null}
+                  {hasSchemes ? (<SchemesContainer />) : null}
+                  {hasSecurityDefinitions ? (<AuthorizeBtnContainer />) : null}
                 </Col>
               </div>
-            </div>
+            ) : null}
 
             <FilterContainer/>
 
