@@ -17,11 +17,19 @@ export default class ModelExample extends React.Component {
     super(props, context)
     let { getConfigs, isExecute } = this.props
     let { defaultModelRendering } = getConfigs()
+
+    let activeTab = defaultModelRendering
+
     if (defaultModelRendering !== "example" && defaultModelRendering !== "model") {
-      defaultModelRendering = "example"
+      activeTab = "example"
     }
+
+    if(isExecute) {
+      activeTab = "example"
+    }
+
     this.state = {
-      activeTab: isExecute ? "example" : defaultModelRendering
+      activeTab: activeTab
     }
   }
 
@@ -33,8 +41,12 @@ export default class ModelExample extends React.Component {
     })
   }
 
-  componentWillReceiveProps(props) {
-    if (props.isExecute && props.isExecute !== this.props.isExecute) {
+  componentWillReceiveProps(nextProps) {
+    if (
+      nextProps.isExecute &&
+      !this.props.isExecute &&
+      this.props.example
+    ) {
       this.setState({ activeTab: "example" })
     }
   }
@@ -43,19 +55,28 @@ export default class ModelExample extends React.Component {
     let { getComponent, specSelectors, schema, example, isExecute, getConfigs, specPath } = this.props
     let { defaultModelExpandDepth } = getConfigs()
     const ModelWrapper = getComponent("ModelWrapper")
+    const HighlightCode = getComponent("highlightCode")
 
-    return <div>
+    let isOAS3 = specSelectors.isOAS3()
+
+    return <div className="model-example">
       <ul className="tab">
         <li className={ "tabitem" + ( this.state.activeTab === "example" ? " active" : "") }>
           <a className="tablinks" data-name="example" onClick={ this.activeTab }>{isExecute ? "Edit Value" : "Example Value"}</a>
         </li>
         { schema ? <li className={ "tabitem" + ( this.state.activeTab === "model" ? " active" : "") }>
-          <a className={ "tablinks" + ( isExecute ? " inactive" : "" )} data-name="model" onClick={ this.activeTab }>Model</a>
+          <a className={ "tablinks" + ( isExecute ? " inactive" : "" )} data-name="model" onClick={ this.activeTab }>
+            {isOAS3 ? "Schema" : "Model" }
+          </a>
         </li> : null }
       </ul>
       <div>
         {
-          this.state.activeTab === "example" && example
+          this.state.activeTab === "example" ? (
+            example ? example : (
+              <HighlightCode value="(no example available)" />
+            )
+          ) : null
         }
         {
           this.state.activeTab === "model" && <ModelWrapper schema={ schema }
