@@ -23,7 +23,7 @@ export default class Model extends ImmutablePureComponent {
       return ref.replace(/^.*#\/definitions\//, "")
     }
     if ( ref.indexOf("#/components/schemas/") !== -1 ) {
-      return ref.replace("#/components/schemas/", "")
+      return ref.replace(/^.*#\/components\/schemas\//, "")
     }
   }
 
