@@ -79,8 +79,8 @@ export default class ObjectModel extends Component {
             {
               <table className="model"><tbody>
               {
-                !description ? null : <tr style={{ color: "#666", fontStyle: "italic" }}>
-                    <td>description:</td>
+                !description ? null : <tr style={{ color: "#666", fontWeight: "normal" }}>
+                    <td style={{ fontWeight: "bold" }}>description:</td>
                     <td>
                       <Markdown source={ description } />
                     </td>
