@@ -6,10 +6,14 @@ import { OAS3ComponentWrapFactory } from "../helpers"
 import { sanitizer } from "core/components/providers/markdown"
 
 const parser = new Remarkable("commonmark")
-
+parser.block.ruler.enable(["table"])
 parser.set({ linkTarget: "_blank" })
 
 export const Markdown = ({ source, className = "" }) => {
+  if(typeof source !== "string") {
+    return null
+  }
+  
   if ( source ) {
     const html = parser.render(source)
     const sanitized = sanitizer(html)
