@@ -2,7 +2,7 @@ import React, { cloneElement } from "react"
 import PropTypes from "prop-types"
 
 //import "./topbar.less"
-import Logo from "./logo_small.png"
+import Logo from "./logo_small.svg"
 import {parseSearch, serializeSearch} from "../../core/utils"
 
 export default class Topbar extends React.Component {
@@ -74,7 +74,7 @@ export default class Topbar extends React.Component {
     const urls = configs.urls || []
 
     if(urls && urls.length) {
-      this.loadSpec(urls[this.state.selectedIndex].url)
+      var targetIndex = this.state.selectedIndex
       let primaryName = configs["urls.primaryName"]
       if(primaryName)
       {
@@ -82,9 +82,12 @@ export default class Topbar extends React.Component {
           if(spec.name === primaryName)
             {
               this.setState({selectedIndex: i})
+              targetIndex = i
             }
         })
       }
+
+      this.loadSpec(urls[targetIndex].url)
     }
   }
 
@@ -116,7 +119,7 @@ export default class Topbar extends React.Component {
       })
 
       control.push(
-        <label className="select-label" htmlFor="select"><span>Select a spec</span>
+        <label className="select-label" htmlFor="select"><span>Select a definition</span>
           <select id="select" disabled={isLoading} onChange={ this.onUrlSelect } value={urls[this.state.selectedIndex].url}>
             {rows}
           </select>
@@ -134,8 +137,7 @@ export default class Topbar extends React.Component {
         <div className="wrapper">
           <div className="topbar-wrapper">
             <Link>
-              <img height="30" width="30" src={ Logo } alt="Swagger UI"/>
-              <span>swagger</span>
+              <img height="40" src={ Logo } alt="Swagger UI"/>
             </Link>
             <form className="download-url-wrapper" onSubmit={formOnSubmit}>
               {control.map((el, i) => cloneElement(el, { key: i }))}
